Remove the @alephium/cli dependency from the generated code
