+++ packages/cli/src/deployment.test.ts
@@ -18,9 +18,8 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 
 import { PrivateKeyWallet } from '@alephium/web3-wallet'
 import { Deployments, DeploymentsPerAddress, recordEqual, validatePrivateKeys } from './deployment'
-import { ContractInstance, web3 } from '@alephium/web3'
+import { DeployContractExecutionResult, ContractInstance, web3 } from '@alephium/web3'
 import { randomContractAddress, testAddress } from '@alephium/web3-test'
-import { DeployContractExecutionResult } from './types'
 
 describe('deployments', () => {
   it('test record equal', () => {
