+++ packages/web3/src/contract/deployment.ts
@@ -16,8 +16,25 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-export * from './ralph'
-export * from './contract'
-export * from './events'
-export * from './script-simulator'
-export * from './deployment'
+import { ContractInstance } from './contract'
+
+export interface ExecutionResult {
+  txId: string
+  unsignedTx: string
+  signature: string
+  gasAmount: number
+  gasPrice: string
+  blockHash: string
+  codeHash: string
+  attoAlphAmount?: string
+  tokens?: Record<string, string>
+}
+
+export interface DeployContractExecutionResult<I extends ContractInstance = ContractInstance> extends ExecutionResult {
+  contractInstance: I
+  issueTokenAmount?: string
+}
+
+export interface RunScriptResult extends ExecutionResult {
+  groupIndex: number
+}
